requirements.txt
pandas==2.2.1
numpy==1.26.4
requests==2.31.0
mysql-connector-python==8.3.0
pycountry==24.6.1
plotly==5.19.0
seaborn==0.13.1
matplotlib==3.8.2
streamlit==1.32.0
